requirements/prod.txt
# anyio==4.6.0
# python-dotenv==1.0.1
pydantic==2.9.2
fastapi==0.115.0
SQLAlchemy==2.0.35
# python-multipart==0.0.12
python-dateutil==2.9.0.post0
uvicorn==0.31.0
alembic==1.13.3
pydantic-settings==2.5.2
asyncpg==0.29.0
beautifulsoup4==4.12.3
httpx==0.27.2
fastapi-utilities==0.2.0